flow-runner/pyproject.toml
[project]
name = "adkflow-runner"
version = "0.1.0"
description = "CLI tool for running ADKFlow workflow YAML files"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.18.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
adkflow = "adkflow.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adkflow"]
